refactor(mesh-v2): extract data-sender mixin
issue #566 の責務分割 6/N。Scratch のグローバル変数を他ノードへ伝播する
ロジックを `data-sender.js` に切り出し。

切り出し対象:
- sendData (キュー時点でのデルタフィルタ + RateLimiter 経由送信)
- _reportData (送信時の最終デルタフィルタ + REPORT_DATA mutation)
- getGlobalVariables / sendAllGlobalVariables
- getRemoteVariable

mesh-service.js: 990 → 884 行 (-106)
data-sender.js: 178 行 (新規)
